# feat: Complete Subscription Feature Modernization - Production Ready