--- a/web/components/subscriptionScheduleList.tsx
+++ b/web/components/subscriptionScheduleList.tsx
@@ -10,7 +10,7 @@ import { Switch } from '@/components/ui/switch'
 interface SubscriptionsListProps {
   subscriptions: ProjectedSubscription[];
   summary: SubscriptionSummary | null;
-  openAddSubscriptionModal: () => void;
+  openAddSubscriptionModal?: () => void;
 }
 
 export function SubscriptionScheduleList({
@@ -104,12 +104,14 @@ export function SubscriptionScheduleList({
       ) : (
         <div className="text-center py-8">
           <p className="text-gray-500">{tSub('noSubscriptions')}</p>
-          <button
-            onClick={openAddSubscriptionModal}
-            className="mt-4 px-4 py-2 bg-black text-white rounded-lg hover:bg-gray-800"
-          >
-            {tSub('addYourFirst')}
-          </button>
+          {openAddSubscriptionModal && (
+            <button
+              onClick={openAddSubscriptionModal}
+              className="mt-4 px-4 py-2 bg-black text-white rounded-lg hover:bg-gray-800"
+            >
+              {tSub('addYourFirst')}
+            </button>
+          )}
         </div>
       )}
     </div>