… web/components/ui/skeletons.tsx
… web/docs/main-page-fab-cleanup.md
… web/docs/subscription-cache-invalidation-verification.md
… web/docs/subscription-cache-test-script.md
… web/docs/subscription-fab-visual-fixes.md
… web/docs/subscription-modal-bug-fixes.md
… web/docs/subscription-schedule-list-cleanup.md
… web/docs/subscriptions-page-modernization-summary.md
… web/docs/subscriptions-page-test-verification.md
… web/docs/subscriptions-view-modernization-summary.md
… web/docs/subscriptions-view-test-verification.md
… web/hooks/queries/useSubscriptionsQuery.ts
… web/stores/appStore.ts